requirements/requirements.txt
absl-py==1.4.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
docutils==0.20.1
easydict==1.10
et-xmlfile==1.1.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
imutils==0.5.4
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.2
keyring==24.2.0
kiwisolver==1.4.4
lazy_loader==0.3
lib==4.0.0
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkginfo==1.9.6
presets==0.1.3
protobuf==4.23.4
pyaml==23.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.1
Pygments==2.15.1
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
pylibjpeg-openjpeg==1.3.2
pyloco==0.0.139
pymedphys-pylibjpeg-libjpeg==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
readme-renderer==40.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
SimpleWebSocketServer==0.1.2
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.7.10
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.65.0
transforms==0.1
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
